Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    int inp,inp1,inp2;
+    float inp,inp1,inp2;
     scanf("%f %f %f", &inp, &inp1, &inp2);
     float avg = (inp + inp1 + inp2) / 3.0;
     printf("Average: %.2f", avg);
